Make parameter serialization consistent with Haskell.

+++ smart-contracts/wasm-chain-integration/src/types.rs
@@ -94,7 +94,7 @@ impl Action {
                 out.extend_from_slice(&(name.len() as u16).to_be_bytes());
                 out.extend_from_slice(&name);
                 out.extend_from_slice(&amount.to_be_bytes());
-                out.extend_from_slice(&(parameter.len() as u32).to_be_bytes());
+                out.extend_from_slice(&(parameter.len() as u16).to_be_bytes());
                 out.extend_from_slice(&parameter);
                 out
             }
